pyproject.toml
[project]
name = "plane-spotter"
version = "0.0.0-alpha0"
description = "Spots planes using ADS-B data and sends events to various backends"
requires-python = ">=3.10"
dependencies = [
    "click",
    "haversine",
    "hydra-core",
    "tweepy",
    "requests",
    "structlog",
    "types-requests",
]

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "freezegun",
    "mypy",
    "pytest",
    "pytest-cov",
    "pytest-httpserver",
    "pytest-structlog",
    "testfixtures",
]

[tool.setuptools]
py-modules = []
include-package-data = true

[tool.setuptools.package-data]
"plane_spotter.data" = ["airport-codes.csv"]


[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
